Make cms_context dynamically allocated.
When we run as a client/server model, we'll want to keep cms_context
around in the daemon, but not have the normal pesign_context around.  So
cms_context becomes our daemon's cryptographic state, and pesign_context
is transient state for a particular task.

With that in mind, we need to have cms_context dynamically allocated.