pesign: Ubuntu build fixes
() pkgconfig already adds the nss and nspr include dirs
to the include paths, and on Ubuntu they are not in
nss3/ or nspr4/ subdirectories of /usr/include; they are
in nss/ and nspr/ instead. Rely on pkgconfig to do the
right thing instead of hard-coding dirs in the source.

() fix link error related to missing pthread symbols

Build-tested on Ubuntu 13.04 and Fedora 18.

Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>