pesign: Add a signing example to pesign man page
I was building linux kernel and it generated an x509 cert and private key
for signing modules. I wanted to use same x509 cert and key to sign bzImage
using pesign. It was not obivious to me how to do that and spent some time
to figure it out.

I think other might run into similar issue. Hence adding an example to
pesign man page.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>