Actually remove all of SpcSpOpusInfo.
So we had SpcSpOpusInfo until c0c4da6c67, when we realized MS no longer
includes those fields in their data, and removed them.

Except it didn't remove the data from the OID table or the enum from the
list, so all the OIDs after that were off by one.

Probably this confirms that a lot of things aren't checking those
completely senseless bits of the Authenticated Attributes.  I can't
decide if that's terrifying or not.

Signed-off-by: Peter Jones <pjones@redhat.com>