Make pesign-authorize-{users,groups} not be in sbin.
If it's in sbin, it needs a man page.  On fedora/rhel it should go in
libexec, on debian probably /usr/share someplace.  This change puts it
in libexec and makes it so you can put that wherever you like by doing:

make libexecdir=/usr/share all
make libexecdir=/usr/share install install_systemd

or similar.

Signed-off-by: Peter Jones <pjones@redhat.com>