Add kernel module signing support
Extend pesign to be able to sign kernel modules in addition to
PE executables.

We use the file extension .ko to detect when to sign it as a module.
The signing involves appending a PKCS7 message with signature, followed
by a Linux-specific module_signature structure.

This initial code supports creating attached and detached signatures,
and supports both standalone pesign and the daemon mode.