Clean up better when generate_spc_signed_data() fails.
When generate_spc_signed_data() fails, such as when unlocking the token
failed, it cleans up the whole arena since the last mark.  That means
cms->ci_digest is freed, but we still have a pointer to it, and
cms_context_fini() is going to try to poison it, which will segfault.

NULL the pointer out so we can avoid that happening.

Signed-off-by: Peter Jones <pjones@redhat.com>