cms_pe_common: Make the opthdr code more transparent to scan-build
scan-build is of the belief that:

  hash_base += pe32opthdr ? sizeof(pe32opthdr->csum)
                          : sizeof(pe64opthdr->csum);

can dereference pe64opthdr, which is null because it was never assigned
to.  It can't, of course, because the case where pe32opthdr /wasn't/
assigned to, pe64opthdr was, and vice versa.  But it can't seem to tell.

Just error check the thing that gets assigned to them and kick out
earlier.

Signed-off-by: Peter Jones <pjones@redhat.com>