Turn off -Wfree-nonheap-object
authvar.c has a call to free (tokenname) where tokenname is set to a string constant
and never changed.  That triggers GCC to issue a diagnostic that the value should not
be passed to free.

This is a false positive from GCC as the call is guarded by a suitable condition that
always happens to be false.  But pesign is being built without optimization and thus
the condition and free call are not optimized away.

This patch just disables the warning.  A better solution would be to fix the sources
or build with the optimizer enabled.