Fix mandoc invocation to not produce garbage
Bizarrely, mandoc doesn't default to outputting man - the default is
"locale", which is either ASCII or UTF-8 (by locale).  This output is
supposed to be some kind of plain-text, but it's formatted so strangely
I'm not sure what the purpose is.  Regardless, it doesn't go well to
feed this into man(1).

Tell mandoc explicitly to produce man pages.

Signed-off-by: Robbie Harwood <rharwood@redhat.com>