read_password(): only prune CR/NL from the end of the file
Right now, when we read the password/PIN from a file, we're pruning the
end of the string from the file we read indiscriminately.  If you don't
have a newline, that means we're cutting off the final digits of the
text.

This changes it to prune only common special characters from the
pinfile, but also to prune /all/ of them.

Related: rhbz#2122777
Signed-off-by: Peter Jones <pjones@redhat.com>