efikeygen: handle tokens that do not support trust
When generating keys on a token efikeygen fails with the following error:
efikeygen.c:add_trust:104: could not set trust for certificate

This happens because CERT_ChangeCertTrust tries to write trust to the
actual token first and then to the internal token that is not
authenticated yet.

Fix this by authenticating internal token and trying again.

Signed-off-by: Peter Jones <pjones@redhat.com>